+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## To Be Released
 
+## 1.39.0
+
 * task(backups-config): update error message when unscheduling periodic backup
 * feat(backups): prevent download of OpenSearch and Elasticsearch backups
 * feat(domains): add 'Manual Action' column if needed